requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.3.0
colorama==0.4.6
fastapi==0.116.2
h11==0.16.0
idna==3.10
joblib==1.5.2
numpy==2.3.3
pandas==2.3.2
pydantic==2.11.9
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.36.0
